# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: File System Recovery & Optimization Simulator API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /filesystem/files:
    post:
      operationId: createFile
      tags: [filesystem]
      summary: Create a new file
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFileBody"
      responses:
        "201":
          description: File created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileNode"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /filesystem/files/{id}:
    delete:
      operationId: deleteFile
      tags: [filesystem]
      summary: Delete a file
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: File deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FilesystemState"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /filesystem/access:
    post:
      operationId: accessFile
      tags: [filesystem]
      summary: Simulate file access (sequential or direct)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccessFileBody"
      responses:
        "200":
          description: Access result with timing metrics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessResult"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    FileNode:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [file, directory]
        size:
          type: integer
        fileType:
          type: string
        createdAt:
          type: string
        parentId:
          type: ["string", "null"]
        blocks:
          type: array
          items:
            type: integer
        status:
          type: string
          enum: [healthy, corrupted, deleted, recovering]
        children:
          type: array
          items:
            $ref: "#/components/schemas/FileNode"
      required: [id, name, type, size, fileType, createdAt, blocks, status]
    FilesystemState:
      type: object
      properties:
        root:
          $ref: "#/components/schemas/FileNode"
        totalBlocks:
          type: integer
        usedBlocks:
          type: integer
        freeBlocks:
          type: integer
        bitmap:
          type: array
          items:
            type: integer
        fragmentationScore:
          type: number
      required: [root, totalBlocks, usedBlocks, freeBlocks, bitmap, fragmentationScore]
...
    CreateFileBody:
      type: object
      properties:
        name:
          type: string
        parentId:
          type: ["string", "null"]
        type:
          type: string
          enum: [file, directory]
        fileType:
          type: string
        size:
          type: integer
      required: [name, type]
    AccessFileBody:
      type: object
      properties:
        fileId:
          type: string
        method:
          type: string
          enum: [sequential, direct]
        operation:
          type: string
          enum: [read, write]
      required: [fileId, method, operation]
    AccessResult:
      type: object
      properties:
        fileId:
          type: string
        method:
          type: string
        operation:
          type: string
        timeTakenMs:
          type: number
        blocksAccessed:
          type: array
          items:
            type: integer
        seekTime:
          type: number
        transferTime:
          type: number
        totalTime:
          type: number
      required: [fileId, method, operation, timeTakenMs, blocksAccessed, seekTime, transferTime, totalTime]
...